Refactor full screen logic to be reusable (#10098)
