+++ .changeset/kind-laws-heal.md
@@ -1,0 +1,10 @@
+---
+"@gradio/annotatedimage": minor
+"@gradio/atoms": minor
+"@gradio/chatbot": minor
+"@gradio/gallery": minor
+"@gradio/image": minor
+"gradio": minor
+---
+
+feat:Refactor full screen logic to be reusable
